pyproject.toml
[tool.poetry]
name = "ai agent from scratch"
version = "0.1.0"
description = "just a quick implementation of ai agent from scratch"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
huggingface-hub = "^0.26.1"
ipykernel = "^6.29.5"
python-dotenv = "^1.0.1"
streamlit = "^1.40.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
